requirements.txt
accelerate==0.27.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.39.0
certifi==2025.10.5
charset-normalizer==3.4.4
datasets==2.16.0
deepspeed==0.10.0
dill==0.3.7
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
filelock==3.20.0
fire==0.7.1
frozenlist==1.8.0
fsspec==2023.10.0
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.8
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonlines==4.0.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmcv==1.7.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
openai==1.57.2
opencv-python==4.10.0.84
packaging==25.0
pandas==2.2.3
peft==0.11.1
pillow==11.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==3.20.0
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.7
pydantic==1.10.24
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==13.9.4
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.21.4
tomli==2.3.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.4
transformers==4.49.0
triton==2.1.0
trl==0.8.6
typeguard==4.4.4
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
urllib3==2.5.0
xformers==0.0.23.post1
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
